+++ src/main/java/com/desi/SmnApp/entities/WeatherStatus.java
@@ -10,7 +10,7 @@ public class WeatherStatus {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	private String status;
 
@@ -18,17 +18,17 @@ public class WeatherStatus {
 		super();
 	}
 
-	public WeatherStatus(int id, String status) {
+	public WeatherStatus(Long id, String status) {
 		super();
 		this.id = id;
 		this.status = status;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
